Mapper cleanups
