+++ libretro/libretro-mapper.h
@@ -1,0 +1,7 @@
+#ifndef LIBRETRO_MAPPER_H
+#define LIBRETRO_MAPPER_H
+
+#define RETRO_DEVICE_VICE_JOYSTICK  RETRO_DEVICE_SUBCLASS(RETRO_DEVICE_JOYPAD, 1)
+#define RETRO_DEVICE_VICE_KEYBOARD  RETRO_DEVICE_SUBCLASS(RETRO_DEVICE_KEYBOARD, 0)
+
+#endif /* LIBRETRO_MAPPER_H */
